This was the already completed code for server
The Server is working fine for just updating the status of the user and signup etc. but it isn't doing the things which we require which is call forwarding to other users as there is no option for taking in the Ip of the user as he goes online and maintain a connection with the user till he is online(we will have to make our android application work in the background so the user doesn't go off every time he exits the application). This Server requires following things
1> Maintain a connection with the client
2> Make sure one user is not using two accounts at one device
3> The server should have a public IP address 
4> The Server can make calls and messages
5> The Application for interacting with the server still has to be made
6> (Optional) Public key Infrastructure (PKI)